changed accounts controller;
 show method where it was @account = Account.find(params[:id]) before. this was allowing access into accounts that users did not own themselves. the new code checks for the ownership of the current user's id against their account id.
owner_id takes in the params that are generated when the link is hit, and the account_variable points to the return value of Account.find(ID).
then if statement checks using the reader method .user_id from the_account if it matches the ID of the current user. once this condition is met the @account gets the found object and all is well.
if not, they are redirected to their accounts page. A future upgrade here can be made wherein the_account will not simply return false throwing a 404, but instead will flash an error.